Change the array type to double
This is a test commit to see if the checker still recognize it as
an array of chars.

+++ 0x01-variables_if_else_while/100-print_comb3.c
@@ -25,7 +25,7 @@ void print_chars(int i, int j, int n)
  */
 int main(void)
 {
-	int x[200];
+	double x[200];
 	int i, j, filled, n1, n2;
 	int state = OUT;
